Moved common covariance components from MF31/MF33/MF40 up one level (Python not ready yet)
