+++ cmake/unit_testing.cmake
@@ -36,6 +36,9 @@ add_subdirectory( src/ENDFtk/record/test )
 add_subdirectory( src/ENDFtk/section/CovariancePairs/test )
 add_subdirectory( src/ENDFtk/section/RectangularMatrix/test )
 add_subdirectory( src/ENDFtk/section/SquareMatrix/test )
+add_subdirectory( src/ENDFtk/section/DerivedRatioToStandard/test )
+add_subdirectory( src/ENDFtk/section/DerivedRedundant/test )
+add_subdirectory( src/ENDFtk/section/ReactionBlock/test )
 add_subdirectory( src/ENDFtk/section/1/451/test )
 add_subdirectory( src/ENDFtk/section/1/452/test )
 add_subdirectory( src/ENDFtk/section/1/455/DecayConstants/test )
@@ -155,9 +158,6 @@ add_subdirectory( src/ENDFtk/section/32/151/UnresolvedRelativeCovariances/JValue
 add_subdirectory( src/ENDFtk/section/32/151/UnresolvedRelativeCovariances/LValue/test )
 add_subdirectory( src/ENDFtk/section/32/151/UnresolvedRelativeCovariances/CovarianceMatrix/test )
 add_subdirectory( src/ENDFtk/section/32/151/UnresolvedRelativeCovariances/test )
-add_subdirectory( src/ENDFtk/section/33/DerivedRatioToStandard/test )
-add_subdirectory( src/ENDFtk/section/33/DerivedRedundant/test )
-add_subdirectory( src/ENDFtk/section/33/ReactionBlock/test )
 add_subdirectory( src/ENDFtk/section/33/test )
 add_subdirectory( src/ENDFtk/section/34/LegendreBlock/test )
 add_subdirectory( src/ENDFtk/section/34/ReactionBlock/test )
