+++ src/ENDFtk/section/31.hpp
@@ -33,23 +33,6 @@ namespace section{
   template<>
   class Type< 31 > : protected Base {
 
-  public:
-
-    // all includes intentionally taken from MF33
-
-    // include classes for subsubsections
-    #include "ENDFtk/section/33/DerivedRedundant.hpp"
-    #include "ENDFtk/section/33/DerivedRatioToStandard.hpp"
-    using DerivedCovariance = std::variant<
-      DerivedRedundant,
-      DerivedRatioToStandard >;
-
-    // include classes for subsections
-    #include "ENDFtk/section/33/ReactionBlock.hpp"
-
-
-  private:
-
     /* fields */
     std::vector< ReactionBlock > reactions_;
