+++ src/ENDFtk/section/DerivedRatioToStandard/src/ctor.hpp
