+++ src/ENDFtk/section/DerivedRatioToStandard/test/CMakeLists.txt
@@ -1,6 +1,6 @@
 
-add_executable( ENDFtk.section.33.ReactionBlock.test ReactionBlock.test.cpp )
-target_compile_options( ENDFtk.section.33.ReactionBlock.test PRIVATE ${${PREFIX}_common_flags}
+add_executable( ENDFtk.section.DerivedRatioToStandard.test DerivedRatioToStandard.test.cpp )
+target_compile_options( ENDFtk.section.DerivedRatioToStandard.test PRIVATE ${${PREFIX}_common_flags}
 $<$<BOOL:${strict}>:${${PREFIX}_strict_flags}>$<$<CONFIG:DEBUG>:
 ${${PREFIX}_DEBUG_flags}
 $<$<BOOL:${coverage}>:${${PREFIX}_coverage_flags}>>
@@ -10,5 +10,5 @@ $<$<BOOL:${link_time_optimization}>:${${PREFIX}_link_time_optimization_flags}>
 $<$<BOOL:${nonportable_optimization}>:${${PREFIX}_nonportable_optimization_flags}>>
 
 ${CXX_appended_flags} ${ENDFtk_appended_flags} )
-target_link_libraries( ENDFtk.section.33.ReactionBlock.test PUBLIC ENDFtk )
-add_test( NAME ENDFtk.section.33.ReactionBlock COMMAND ENDFtk.section.33.ReactionBlock.test )
+target_link_libraries( ENDFtk.section.DerivedRatioToStandard.test PUBLIC ENDFtk )
+add_test( NAME ENDFtk.section.DerivedRatioToStandard COMMAND ENDFtk.section.DerivedRatioToStandard.test )
